backend/requirements.txt
# Server
firebase-admin
flask
flask-cors
google-cloud-firestore
google-cloud-texttospeech
google-cloud-storage
google-cloud-secret-manager
gunicorn
openai
requests
# cython

# Wav2Lip
# librosa==0.7.0
# numpy==1.17.1
opencv-contrib-python>=4.2.0.34
opencv-python
# torch==1.1.0
# torchvision==0.3.0
# tqdm==4.45.0
# numba==0.48

# RTVC
umap-learn
visdom
librosa>=0.8.0
matplotlib>=3.3.0
# numpy==1.19.3; platform_system == "Windows"
numpy==1.19.4
# scipy>=1.0.0
tqdm
sounddevice
SoundFile
Unidecode
inflect
PyQt5
multiprocess
numba
webrtcvad; platform_system != "Windows"

# First-Order Model
imageio==2.3.0
# matplotlib==2.2.2
# numpy==1.15.0
# pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.1
scikit-learn==0.19.2
scikit-image==0.14.0
# scipy==1.1.0
# torch==1.0.0
# torchvision==0.2.1
# tqdm==4.24.0
